Move GitOid code and tests into new crate.
This commit relocates the `GitOid`, `HashAlgorithm`, and `Source`
types, along with their associated trait impls, methods, constants,
and tests, to the new `gitoid` crate. It also updates the `gitbom`
crate to only include the `gitbom`-specific tests, and to depend on
the `gitoid` crate (currently by a path dependency with no version
specified) for providing the necessary types.

Signed-off-by: Andrew Lilley Brinker <alilleybrinker@gmail.com>