Improve consistency of gitoid tests.
This commit updates the `tests` module in the `gitoid` crate to
make sure that tests between SHA1 and SHA256 and between byte
inputs and buffer inputs are consistent. It also moves tests which
previously used `unwrap` to instead use `?`, with the intent of
reducing line noise.

Note that some tests now include more than one `assert_eq`, to
ensure both that the hashing is correct (this is really just
checking that the correct hashing method is called --- we assume
that the dependencies used for hashing implement the hash functions
correctly), and that the construction of the `gitoid` string works
as-intended.

Signed-off-by: Andrew Lilley Brinker <alilleybrinker@gmail.com>