Rename GitOid::new to GitOid::new_from_bytes.
The major change of this commit is in the title: renaming
`GitOid::new` to `GitOid::new_from_bytes`. This avoids placing
constructor-supremacy on the from-bytes version, as there are
equally-valid constructors working from strings or from arbitrary
readers.

This commit also reorganizes the methods in the `gitoid` module.
First by collecting all the constructors at the start of the impl
block for the `GitOid` struct, and then by gathering the getters
for the results of `GitOid` construction.

Finally, the helper methods for actually performing the hashing
from whatever input source is used are moved out of the `GitOid`
struct and renamed to be clearer that they are producing bytes
from either an arbitrary set of readers or an arbitrary set of
async readers.

Signed-off-by: Andrew Lilley Brinker <alilleybrinker@gmail.com>