Revert changes to `gitbom` tests.
This commit reverts changes to the `gitbom` tests so they'll pass.
The right dance in the future when `gitoid` makes a breaking change
will be to publish a breaking version of it to crates.io, and only
after that make the changes to bring `gitbom` up to use the new
version.

Signed-off-by: Andrew Lilley Brinker <alilleybrinker@gmail.com>