Introduce a separate Hash struct.
This is intended to make the API clearer in separating the hash,
which is part of the `GitOid` but not sufficient for representing
the `GitOid`, from the overall `GitOid` structure. The `GitOid`
can be represented as a URI, or you can extract the parts (hash
algorithm, object type, hash) individually. If you have the `Hash`,
now represented as a struct, you can either access the individual
bytes through the `Deref` impl or `as_bytes` method, or you can
get the hexadecimal encoding through the `Display` impl or the
implied `to_string` method or explicit `hex` method.

Signed-off-by: Andrew Lilley Brinker <alilleybrinker@gmail.com>