Add `GitOid::new_from_url` and `TryFrom`.
This commit adds the ability to try to construct a `GitOid` from
a `url::Url` by parsing the parts of the `Url` to ensure they
contain all the required data, and then constructing a `GitOid`
if they do.

This includes additions to enable parsing `ObjectType` and
`HashAlgorithm` from a string slice, plus new error variants
(with either `String` or `Box<Url>` fields) to represent different
modes of failure for parsing.

This commit also renames `GitOid::uri` to `GitOid::url`.

The new name is a more-specific explanation of what we're dealing
with, and matches the type returned (`url::Url`). This change
also makes `GitOid::url` consistent with `GitOid::new_from_url`.
I also added new tests to validate that the correct errors are
produced for different types of invalid `Url`s.

Signed-off-by: Andrew Lilley Brinker <alilleybrinker@gmail.com>