First draft of a finder implementation
This commits adds a new `Finder` struct which matches different
sources of `GitOid`s against a known set of `GitOid`s, returning
both an identifier and the matching `GitOid` itself, along with
reporting any errors that arise when resolving the new `GitOid`s.

The docs for the `Finder` struct spell it out the clearest. This
design is centered around the new `IntoIdentifiedGitOid` trait,
which is implemented for a variety of existing types. The goal of
this is to allow the `Finder` to be generic over things which
produce `GitOid`s, so that, for example, someone could add a
mechanism to resolve `GitOid`s from a `Url` without changing any
existing code.

Signed-off-by: Andrew Lilley Brinker <alilleybrinker@gmail.com>