feat: Make GitOid::new_invalid pub(crate)
For the C API to handle errors, we have a constructor for GitOid
called new_invalid that produces an _invalid_, _unusable_ GitOid.
Specifically, this bad GitOid has a length of 0, and the hash value
is a zero-initialized array. When FFI functions expected to return a
GitOid fail, they can return in invalid GitOid instead, and we also
expose an FFI function for that GitOid.

However, this is all specific to the C FFI, and only needs to be called
by our own FFI functions when they fail. It should _never_ be called
directly by any Rust or C users of the GitOid crate, so we can mark it
pub(crate) to avoid exposing the symbol to the linker.