fix: Add missing conditional compilation for unix imports
The `ffi/gitoid.rs` file included some `std::os::unix` imports unconditionally,
which wasn't caught in CI but causes builds of the crate to fail on
non-Unix systems. This is a limitation of our CI today, and in the
future it would be great to have WIndows and other CI platforms
set up so we can catch these issues pre-merge.

This commit adds conditional compilation annotations so the imports
are only done on the appropriate platform.