feat: Improve clarity / reliability of FFI tests
We use a C program to validate the FFI, and the previous version
did an odd thing for testing `gitoid_new_from_bytes`, where it used
`sscanf` through a string into a buffer. This actually did't appear to
workas intended, and I've opted to replace it with a literal buffer
containing some bytes instead.

I also took the opportunity to add more validity checks to some other
tests, and to simplify construction of the test array by introducing a
new TEST macro.

Signed-off-by: Andrew Lilley Brinker <abrinker@mitre.org>