feat: Bump dependency versions to latest.
This commit bumps the sha1, sha2, and url dependencies to
their latest versions. For sha1 and sha2 this is just a patch version
bump. For url it's a minor bump. Nothing broke with these updates,
and in generally Cargo would select the latest compatible anyway.
This is just good for clarity to the resolver.

Signed-off-by: Andrew Lilley Brinker <abrinker@mitre.org>