fix: Remove unused dependencies.
The `omnibor` crate doesn't actually use the hex, im, sha1,
or sha2 crates anymore, and doesn't need to. All of that
functionality is now in the `gitoid` crate that it depends on,
so these deps can be removed.

Note that the `omnibor` crate still fails to compile on
Windows due to a prior issue, which can be resolved
(and the deps version bumped) with the publication of
a new version of the `gitoid` crate.

Signed-off-by: Andrew Lilley Brinker <abrinker@mitre.org>