fix: Fix broken compilation of `omnibor` crate.
Yes, this "fixes" everything by deleting the (minimal) contents of
the crate. However, the existing contents of the crate...

1. Do not do anything meaningful.
2. Overlap in a confusing way with the much-more-developed
   `gitoid` crate.
3. Are set to be rewritten to reflect updates to the OmniBOR
   specification.

I personally feel it's better to be real about the state of this
crate, which is that it is not meaningfully useful in the new
OR old form, and move on from there, rather than publish an API
that isn't meaningfully useful and is also broken.

Signed-off-by: Andrew Lilley Brinker <abrinker@mitre.org>