Make `GitOid` generic over the underlying hash output size.
This commit redesigns the implementation of the `GitOid` type,
changing the underlying buffer from one designed to fit the largest
possible hash algorithm _supported_ by the crate regardless of the
actual hash algorithm used in a specific case, to one that is sized
to have exactly the space needed for the selected hash algorithm.

This involves a little bit more complexity than I'd like, because Rust's
const-generics support is still incomplete and thus this crate
relies on the `generic_array` crate to provide what we need. In the
future we'd replace this with normal const-generic arrays which
would (hopefully) also simplify the bounds on the generic types.

This commit does not complete the work for this transition, as it
does not address fixing the FFI story. The FFI module is commented
out, and will need to be updated before this can be merged.

This commit _does_ remove the finder API. After consideration,
I think the current API for finding matching GitOids is both too
complicated in terms of trait bounds AND under-powered in terms
of controlling matching behavior. It may not even make sense to
proscribe _how_ matching can be performed on a scale beyond
"check object against GitOid" given the complexities of performance
tuning which are possible.

A fun note on sizing, in the prior code the size of a GitOid was fairly
large, as the GitOid type had to store two 4 byte enums (HashAlgorithm
and ObjectType) plus a "usize" (usually 4 bytes), and a 32-byte buffer.
So it would generally be 52 bytes.

In the new code, for SHA-256, GitOids are 33 bytes (1 byte for the
ObjectType, and 32 bytes for the buffer), and for SHA-1 GitOids are
21 bytes (1 byte for the ObjectType, 20 bytes for the buffer). Given that
we expect code using GitOids in OmniBOR to be using large numbers of
them, this size saving is great news!

Signed-off-by: Andrew Lilley Brinker <alilleybrinker@gmail.com>