Update FFI for new GitOid type
This transitions the FFI to an opaque-pointer approach
for the GitOid types since they're now generic and not
exposable to C code directly.

It uses a macro to handle generating functions for each
of the concrete GitOid types depending on the hash
algoirthm used, which means the C FFI has to be
built using a Nightly compiler so cbindgen can
expand the macros with an unstable rustc flag.

Signed-off-by: Andrew Lilley Brinker <alilleybrinker@gmail.com>