feat: rewrite docs intro, reorg docs slightly (#94)
This commit rewrites the introductory docs for the crate,
and also reorganizes some docs for FFI and hashes slightly
to be clearer / cleaner. FFI is no longer marked doc(hidden),
and the contents of crate::ffi::gitoid have been re-exported
to appear under crate::ffi directly. This isn't _really_ that
important since they're only intended to be called from C
code anyway, but at least this way they appear in the Rust
docs and are easier to find.

Signed-off-by: Andrew Lilley Brinker <alilleybrinker@gmail.com>