feat: first draft of README rewrite (#88)
This commit rewrites the README to more clearly explain
the background and goals of OmniBOR and of the Rust
implementation.

Right now it delegates info to crate-specific
READMEs which haven't been written yet. So there's more
work to do prior to merge. We also
need to at least publish a placeholder omnibor crate to crates.io.

Signed-off-by: Andrew Lilley Brinker <alilleybrinker@gmail.com>