fix: resolved transient cbindgen issue
This commit fixes a transient issue where cbindgen would sometimes pick
up and try to build gitoid:0.3.0 instead of the current version
gitoid:0.4.0. This was happening because the omnibor crate imported
gitoid:0.3.0, and the two names would both appear in the output of
`cargo metadata`. cbindgen would use the first one that appeared, which
was sometimes gitoid:0.4.0 and sometimes gitoid:0.3.0. If it was the
older version, then everything would break.

This isn't really a long term fix, so there's a need to do more in the
future to stop this from happening again. For now this is enough.

Signed-off-by: Andrew Lilley Brinker <alilleybrinker@gmail.com>