feat: Windows test, FFI test, and commit msg CI (#106)
* feat: Windows test, FFI test, and commit msg CI

This commit improves the CI configuration by:

- Adding testing for Windows
- Adding testing for CCI code on Ubuntu and Windows
- Adding checking for conventional commit format in commit messages
- Upgrading from checkout v2 to v4 across the board.

Signed-off-by: Andrew Lilley Brinker <alilleybrinker@gmail.com>

* fix: resolved transient cbindgen issue

This commit fixes a transient issue where cbindgen would sometimes pick
up and try to build gitoid:0.3.0 instead of the current version
gitoid:0.4.0. This was happening because the omnibor crate imported
gitoid:0.3.0, and the two names would both appear in the output of
`cargo metadata`. cbindgen would use the first one that appeared, which
was sometimes gitoid:0.4.0 and sometimes gitoid:0.3.0. If it was the
older version, then everything would break.

This isn't really a long term fix, so there's a need to do more in the
future to stop this from happening again. For now this is enough.

Signed-off-by: Andrew Lilley Brinker <alilleybrinker@gmail.com>

* fix: Update CI jobs from make to shell script.

Haha forgot to do this when I made the replacement. We weren't really
using the features of Make so I thought it would be better to just
use a shell script. This commit updates CI to use that script.

Signed-off-by: Andrew Lilley Brinker <alilleybrinker@gmail.com>

---------

Signed-off-by: Andrew Lilley Brinker <alilleybrinker@gmail.com>