feat: simplify GitOid crate substantially. (#108)
This commit comments out the FFI for the GitOid crate, which
will be repurposed into an FFI for the OmniBOR crate. This will avoid
a known issue with cbindgen involving macro expansion of crates used
as dependencies in the same workspace, and will also more directly
reflect our intended use of the code in this workspace anyway. We want
to provide a basis for cross-language OmniBOR implementations, not
really for cross-language GitOID implementations, for which libgit2
should already be fine.

This commit also simplifies the trait bounds on the GitOID type, and
modifies some existing imports from `std` to `core`.

Signed-off-by: Andrew Lilley Brinker <alilleybrinker@gmail.com>