feat: Add async constructors for GitOid.
This commits adds two new constructors for `GitOid` to build new
GitOIDs from async readers, with or without an expected length
being provided.

This was pretty easy! I've generally stayed away from doing async
things because of the general vibe that it's kinda finicky / difficult.
I'm happy this was so simple to do.

Signed-off-by: Andrew Lilley Brinker <alilleybrinker@gmail.com>