feat: improve GitOid trait bounds (#111)
This commit fully hides the implementation of the hash algorithms
by augmenting the HashAlgorithm trait to hide the digester types
inside of it. This makes for some slightly more complicated
internal code, but a much nicer external interface. It also means we
no longer need to re-export the cryptographic crates we use.

Signed-off-by: Andrew Lilley Brinker <alilleybrinker@gmail.com>