feat: Minor cleanup of docs and trait defs. (#112)
This commit marks the contents of the HashAlgorithm and ObjectType
traits as hidden, because their internals are _only_ relevant to the
internals of `GitOid`. The trait is sealed, so it can't be implemented
by other types, and the types that _do_ implement it aren't able to be
constructed (no public constructor + private field).

This also moves all exported items to the top level, so there are no
public submodules. The API of the crate is quite small, so it's nice
for users to have all the docs accessible from one place.

Signed-off-by: Andrew Lilley Brinker <alilleybrinker@gmail.com>