feat: introduce new OmniBOR CLI.
This commit introduces a new, and currently very basic, OmniBOR CLI.

At the moment, this CLI is only able to produce Artifact Identifiers
for files it's given. In the future it'll be expanded with more
functionality.

To use the CLI during development, run:

    cargo run --features="build-binary" -- <CLI ARGS>

Replace `<CLI ARGS>` with arguments for the CLI. Use the `--help` flag
to find out the API of the CLI.

Signed-off-by: Andrew Lilley Brinker <alilleybrinker@gmail.com>