feat: add --hash flag to CLI commands
This commit adds a --hash/-H flag to the CLI commands to support
specifying the underlying hash algorithm to use for IDs. Currently
only SHA-256 is supposed, but the OmniBOR spec is designed to handle
future changes to add other algorithms (for example, if SHA-256
is later broken and collisions become possible to generate).

It also mildly improves help text for the CLI.

Signed-off-by: Andrew Lilley Brinker <alilleybrinker@gmail.com>