docs: Add gitoid crate README, update crate desc.
This commit includes a README for the `gitoid` crate which documents a
bit about what the crate is for and how it's designed. There's more to
add in the future, but it's a starting point.

This commit also replaces the current description of the `gitoid` crate
with a better one that should hopefully be clearer to users on
Crates.io.

Signed-off-by: Andrew Lilley Brinker <alilleybrinker@gmail.com>