feat: Add 'cargo xtask' for custom tasks
This commit implements the initial scaffolding of the 'cargo xtask'
pattern for the OmniBOR Rust workspace. In particular, it defines
logging, a basic CLI, a mechanism for pipelining and rollback of steps
in a process, and the rough outline of a "release" subcommand which
will in the future handle releasing new versions of crates in the
workspace.

Signed-off-by: Andrew Lilley Brinker <alilleybrinker@gmail.com>