chore: Split up CLI crate into modules.
Previously, all the logic of the CLI crate was together in a single
file, but this was getting increasingly hard to manage. This commit
breaks up that CLI into a bunch of separate modules.

It also mildly refactors the print queue logic to be behind a Printer
type, which simplifies use a little bit.

Signed-off-by: Andrew Lilley Brinker <alilleybrinker@gmail.com>