fix: Fix 'find' command short format.
Previously, the 'short' format on the 'find' command would actually
just re-print the hash that was provided, as it was treating the
output identically to the `id` command. This commit introduces a new
`Msg` constructor for the `find` version, which is identical to the
`id` version for the `plain` and `json` formats, but which _only_
prints the found path in the 'short' format.

It also slightly refactors printing in `Msg` to avoid the trait object
and instead just match on the message status to decide where to write.

Signed-off-by: GitHub <noreply@github.com>