feat: Progress implementing `release` for `xtask` (#151)
This commit moves closer to a full implementation of the `release`
subcommand for `cargo xtask`. This includes some improvements to the
underlying pipeline-handling code, as well as improvements to the
release workflow itself. There are now more checks run before
executing anything, as well as guaranteed rollback, and implementation
of the changelog generation is underway.

Signed-off-by: Andrew Lilley Brinker <alilleybrinker@gmail.com>