feat(xtask): Check `git-cliff` version bump
This commit adds another step to the `cargo xtask release` flow to
check the version bump auto-calculated by `git-cliff` against the one
requested by the user.

The basic idea is that `git-cliff` calculates the version to bump
automatically based on parsing the prefixes of the commits according
to the Conventional Commits standard and applying some customizable
parsing rules. In our case, the rules are currently the default.

We _could_ just have the whole release go by what `git-cliff` suggests,
but it's probably better to have the user provide their expected
version bump and then have `git-cliff` check against that, and error
out if they're different.

Signed-off-by: Andrew Lilley Brinker <alilleybrinker@gmail.com>