feat(xtask): Implemented final 'release' step.
This commit adds the final step of the `cargo xtask release` subcommand:
actually releasing the crate! Everything prior involves validating the
environment or checking the changelog, but the final step actually
publishes a new version. It still needs to be tested from a deployment
machine, but this should be pretty close!

Signed-off-by: Andrew Lilley Brinker <alilleybrinker@gmail.com>