fix(xtask): Correct version prefix to include 'v'
For the version comparison against what's expected by `git-cliff`, we
have to do some version string parsing, and this involves stripping
a prefix off which specifies the crate to _only_ get the actual
version number triple portion. The prior prefix used was missing the
letter "v" which is part of the prefix, and so parsing failed in
practice. This commit adds the missing part of the prefix.

Signed-off-by: Andrew Lilley Brinker <alilleybrinker@gmail.com>