fix(xtask): Ignore missing files on rollback
Previously, when rolling back the creation of the CHANGELOG file, we
would error out if the file couldn't be found for deletion. However,
if the commitment step afterward runs and gets rolled back, it will
un-commit and thereby delete the file anyway, so we can instead use
`rm -f` to _not_ error out if the file has already been deleted.

Signed-off-by: Andrew Lilley Brinker <alilleybrinker@gmail.com>