feat(omnibor): Introduce `omnibor` FFI.
This commit brings over and modifies the prior `gitoid`-crate FFI to
instead expose an `omnibor`-crate FFI. This FFI exposes all the main
functionality of the `ArtifactId` type.

It _does_ currently require an allocator to work, as it allocates the
`ArtifactId`s on the heap and then provides an opaque pointer to the
C cide.

This commit has not yet brought over the `cbindgen` stuff or the
C testing mechanism.

Signed-off-by: Andrew Lilley Brinker <alilleybrinker@gmail.com>