feat(xtask): Make `release` pickup Cargo creds.
Previously, `cargo-release` being run in the sub-shell wasn't picking
up Cargo's authentication information, even if you'd previously run
'cargo login' to authenticate with the Crates.io registry. This commit
is intended to fix that by manually extracting the saved token from
Cargo's credentials file and passing it into the subshell using an
environment variable.

With how `xshell` works, this environment variable won't be logged
anywhere, so it's safe to pass the authentication token this way.

Signed-off-by: Andrew Lilley Brinker <alilleybrinker@gmail.com>