chore: Fix broken `CHANGELOG.md`
The prior initial release of a `CHANGELOG.md` was an experiment with
the new `cargo xtask release` command, and unfortunately I forgot to
include the `--bump` flag to actually have it mark the `UNRELEASED`
items as part of the next version release. This is fixed manually here,
and will be fixed in the automation in a subsequent commit.

Signed-off-by: Andrew Lilley Brinker <alilleybrinker@gmail.com>