feat(gitoid): Rename `GitOid` methods for clarity
This commit modifies the method names for a number of `GitOid` methods
to clarify when the thing being passed in is being _hashed_ (any of the
`id_*` methods) vs. when it's being used as an identity itself to
construct a `GitOid` from `try_from_url`.

This should hopefully make it clearer how to _use_ the `GitOid` type
correctly!

This commit also modifies the `FromStr` implementation to use
`try_from_url` under the hood instead of `id_str`, as it seems
surprising to think that `FromStr` would actually be hashing the
contents to produce an identifier rather than being a convenient
shorthand for first parsing and then constructing from a
`gitoid`-scheme URL.

Signed-off-by: Andrew Lilley Brinker <alilleybrinker@gmail.com>