fix(xtask): Add back missing `release` steps
Looks like I'd accidentally commented out and then committed the
removal of two release steps. Thankfully it was inconsequential in
terms of accidentally permitting mistaken releases, but we do want them
turned back on.

This commit also cleans up the configuration of the workspace and of
`cargo release` specifically to produce better commit messages when
releasing a new version.

Signed-off-by: Andrew Lilley Brinker <alilleybrinker@gmail.com>