feat(cli): Split out CLI to its own package. (#171)
This commit moves the existing CLI implementation from
`omnibor/src/bin/omnibor` to its own top-level package called
`omnibor-cli`. This includes modifying the package manifest for the
`omnibor` package to no longer refer to the binary crate within it,
updating the dependencies for both `omnibor` and `omnibor-cli` to
reflect only the minimally-necessary for each respectively, and
configuring `omnibor-cli` to still produce a binary called `omnibor`,
and to be the default crate build when using `cargo run` within the
workspace without specifying a specific binary crate.

This is basically motivated by the difficulties of having a library
and binary crate within a single package. In theory, it's supported
by Cargo, but in practice it's difficult to work with and honestly
more trouble than it's worth.

This also means `omnibor-cli` can be versioned separately from
`omnibor`, though for consistency I'm starting it out on its "current"
version tied to the `omnibor` crate (`0.5.1`). In the future these can
diverge without issue.

This also means `omnibor-cli` can have its own changelog, and that its
deployment with `cargo-dist` can be simplified. Plus, we don't have
to deal with the annoying `default-features` nonsense in Cargo
anymore.

It should also make contributing to the CLI easier by no longer
burying the source for it deep within the source tree of the `omnibor`
crate.

Lots of benefits for doing it this way!

Signed-off-by: Andrew Lilley Brinker <alilleybrinker@gmail.com>