fix(xtask): Fix broken version parsing in release
This commit updates how version parsing is done in
`cargo xtask release` to properly handle the new `omnibor-cli` crate.
It also adds an `--allow-dirty` flag to make it possible to iterate
on debugging the `release` task without having to first commit every
little edit. This flag, when active, just causes the `check-git-ready`
step to log a warning instead of erroring out if Git's worktree or
staging area are dirty.

Signed-off-by: Andrew Lilley Brinker <alilleybrinker@gmail.com>