fix: make `git-cliff` simplify version numbers
Previously, the `git-cliff` template we had configured would generate
section headings based solely on the Git tags picked up for new
versions, which were prefixed with the name of the crate being
released by `cargo-release`. These headers looked like `gitoid-v0.5.1`,
and unfortunately would _not_ be picked up by `cargo-dist` when
generating GitHub releases. This commit modifies the template to put
the headers in a simpler format that `cargo-dist` _should_ be able
to discover.

Signed-off-by: Andrew Lilley Brinker <alilleybrinker@gmail.com>