feat(omnibor): First pass at `InputManifest`
This commit is a first-pass implementation of `InputManifest`s, without
any logic for actually constructing them. My intent is that they'll be
made by a builder type, and be immutable once constructed. To that end,
all fields are private and gettable only by accessors (which should
generally be inlined to callers to avoid a perf hit).

This commit also implements `Display` to get the OmniBOR-spec-compliant
printed version of a manifest. This included some light
Clippy-wrangling to ensure we only do `\n` newlines.

The next step will probably be to implement the builders.

Signed-off-by: Andrew Lilley Brinker <alilleybrinker@gmail.com>