feat(omnibor): Enrich relations in `InputManifest`
This commit enriches relation information in `InputManifest`, encoding
additional info about how the related artifact is related to the
target artifact. This is currently non-standard in the OmniBOR
specification, though it has been discussed among the OmniBOR Working
Group informally.

This commit also explicitly stores the target artifact (the artifact
an `InputManifest` is describing) in the manifest data. When manifests
are persisted this will be used to either drive embedding or drive
some non-embedding mechanism for relating the two.

This commit also eliminates the `Display` impl for the relation type,
making it instead so that _only_ the `InputManifest` has a `Display`
impl, as it doesn't make sense to encode relation rows individually.

Signed-off-by: Andrew Lilley Brinker <alilleybrinker@gmail.com>