feat(omnibor): Add missing `Relation::kind` method
This commit adds a `kind` method to `Relation` to enable getting the
relation kind out. This was an oversight, as all fields on the type
are purposefully private to permit future rewriting of the interals
as necessary. This also required adding `Copy` and `Clone` derivations
for `RelationKind`, which is sensible given that it's just a plain enum
without any associated data, and thus sensible and efficient to pass
copies by value rather than handing out references.

Signed-off-by: Andrew Lilley Brinker <alilleybrinker@gmail.com>