feat: added `InputManifest::from_file`
This commit adds the ability to parse an input manifest from a file.

Note that this is not a particularly smart or helpful parser. In the
future I'd definitely want to rewrite this to produce meaningful and
helpful error messages, perhaps using something like `miette`, and
probably leveraging `nom` or some similar parsing crate. This version is
quick and dirty.

This also hasn't really been tested yet, and should be tested before
being merged.

Signed-off-by: Andrew Lilley Brinker <abrinker@mitre.org>