feat(gitoid): Add BoringSSL backend support for SHA-1 and SHA-256 in gitoid
This commit introduces support for the BoringSSL backend for SHA-1 and
SHA-256 in the gitoid Rust package. The SHA-1CD algorithm is not
supported with this backend. To enable BoringSSL, use the "boring"
feature in Cargo.

Additionally, benchmarks have been added to measure the performance
of the implemented cryptographic algorithms.

Signed-off-by: Frederick F. Kautz IV <fkautz@alumni.cmu.edu>