feat: Added `InMemoryStorage`
This commit adds an `InMemoryStorage` type which implements the
`Storage` trait and can be used for testing.

Implementing this type also led to some changes in the design of the
`Storage` trait, including parameterizing the trait over the type of
hash function. This allows storages to advertise what hash functions
they support storing.

Signed-off-by: Andrew Lilley Brinker <alilleybrinker@gmail.com>