feat: Initial implementation of `InputManifest` (#180)
This commit is an initial implementation of `InputManifest`s, including
the ability to construct them with an `InputManifestBuilder`,
to store them in memory with `InMemoryStorage`, and the ability
to store them on disk with `FileSystemStorage`.

It does not include an implementation of embedding mode, where
artifacts themselves are updated to refer to their manifests, though
the code for doing so is at least sketched out and prepared to
be implemented in the future.

It's worth pointing out the few ways this implementation
differs from what's currently in the OmniBOR specification.

First, the directory where manifests are stored by the
`FileSystemStorage` type is `${OMNIBOR_DIR}/manifests` instead
of `${OMNIBOR_DIR}/objects`. I just felt this was a clearer and
more meaningful name.

Second, this implementation tracks a mapping from manifests
to their target artifacts using an "index file" at
`${OMNIBOR_DIR}/targets`. This file is formatted as a space
separated key-value pair per line, where the first value (the key)
is the Artifact ID of the manifest, and the second value (the value)
is the Artifact ID of the target.

Third, we continue to only implement support for `Sha256`-based
Artifact IDs. This is an area of active discussion around the spec
today.

The code as present includes only the most basic tests, though
it does appear to work in practice with common use cases. Further
testing will be necessary.

Signed-off-by: Andrew Lilley Brinker <abrinker@mitre.org>