feat(gitoid): Add BoringSSL backend support for SHA-1 and SHA-256 in gitoid (#182)
* feat(gitoid): Add BoringSSL backend support for SHA-1 and SHA-256 in gitoid

This commit introduces support for the BoringSSL backend for SHA-1 and
SHA-256 in the gitoid Rust package. The SHA-1CD algorithm is not
supported with this backend. To enable BoringSSL, use the "boring"
feature in Cargo.

Additionally, benchmarks have been added to measure the performance
of the implemented cryptographic algorithms.

Signed-off-by: Frederick F. Kautz IV <fkautz@alumni.cmu.edu>

* chore(gitoid): removed some commented unnecessary code

Signed-off-by: Frederick F. Kautz IV <fkautz@alumni.cmu.edu>

* chore(gitoid): Refactor feature setup and conditional compilation in gitoid

- Remove conditional compilation annotations within `boring_sha.rs` since the
  entire module is conditionally compiled.
- Introduce separate features for RustCrypto and BoringSSL backends.
- Adjust `Cargo.toml` to include `rustcrypto` and `boringssl` features.
- Ensure `boringssl` always activates `sha1` and `sha256` since it provides
  both.
- Modify logic to prevent `sha1cd` from being used with only the BoringSSL
  backend.
- Update hash algorithm implementations in `hash_algorithm.rs` to respect the
  new feature setup.
- Add compilation errors for invalid feature combinations to `lib.rs`.
- Update tests in `tests.rs` to conditionally compile based on feature setup.

Signed-off-by: Frederick F. Kautz IV <fkautz@alumni.cmu.edu>

* chore(gh-actions): Update CI workflow to include BoringSSL feature

- Modify `.github/workflows/ci.yml` to build and test with the `boring` feature enabled.

Signed-off-by: Frederick F. Kautz IV <fkautz@alumni.cmu.edu>

* feat: Make "boringssl" and "rustcrypto" features non-exclusive.

This commit makes it so both the "boringssl" and "rustcrypto" features
can be active at the same time, meaning users of the `gitoid` crate will
be able to select at runtime which of the algorithms they want to use,
rather than solely deciding at compile time.

Signed-off-by: Andrew Lilley Brinker <alilleybrinker@gmail.com>

* chore: Hide internal types for "boringssl" backend

This commit hides the `BoringSha1` and `BoringSha256` types from the
`boringssl` backend. They technically have to be public because the
associated type of a public trait must be public, but they only exist to
wrap the types from the `boring` crate and implement the relevant traits
from the `digest` crate so those types can fit into the interface we
require for a cryptography backend. We don't _really_ want downstream
users to use these types, so we make them public but `#[doc(hidden)]`.

Signed-off-by: Andrew Lilley Brinker <alilleybrinker@gmail.com>

* fix(ci): correct build feature from 'boring' to 'boringssl'

Signed-off-by: Frederick F. Kautz IV <fkautz@alumni.cmu.edu>

---------

Signed-off-by: Frederick F. Kautz IV <fkautz@alumni.cmu.edu>
Signed-off-by: Andrew Lilley Brinker <alilleybrinker@gmail.com>
Co-authored-by: Andrew Lilley Brinker <alilleybrinker@gmail.com>