feat: Further simplify feature set for `gitoid`
Further reduce the number of distinct features for the `gitoid` crate by
combining several of the previously-distinct features (`hex`, `serde`, `url`),
which would turn a single dependency on or off, to be part of the `std`
feature. This means a little less feature-flexibility for users, but a much
simpler set of feature combinations.

Since the "backend-" and "hash-" features are intended to be orthogonal, that
basically means we have three "levels" of features in the crate.

- `no_std`
- `std` (`no_std` + standard library features)
- `async` (`no_std` + standard library feates + async features)

`async` implies `std`, so there's no case of `async` without `std`, giving us
this nice feature hierarchy.

Signed-off-by: Andrew Lilley Brinker <alilleybrinker@gmail.com>