feat: Refactoring error types.
This splits the formerly-single `Error` type into two types, `ArtifactIdError`
and `InputManifestError` that mirror the two main types of the crate. I think
this makes the use of the resulting APIs clearer, as the number of possible
error cases when dealing with Artifact IDs is much smaller than when dealing
with Input Manifests.

Signed-off-by: Andrew Lilley Brinker <alilleybrinker@gmail.com>