chore: More documentation.
This updates the main documentation page with a list of goals to complete
before releasing version 1.0.0 of this crate.

It's great to think we're getting closer and more stable! I think embedding
is the one major technical hurdle remaining, and besides that it's a number of
smaller mechanical things to fully expose our API to other languages via FFI,
to enable complete cryptography library flexibility, and to fully document
the APIs in the crate.

Signed-off-by: Andrew Lilley Brinker <alilleybrinker@gmail.com>