feat(cli)!: Make `manifest create` input files positional
This amends the `manifest create` arguments, turning the list of input files
from a flag (`-i` / `--input`) into a positional argument. Normally I'm wary of
positional arguments because they're less backwards compatible [1], but in this
case this argument should always be the primary one, and making it positional
makes it easier to integrate with things like shell subcommands to produce a
file list.

[1]: https://clig.dev/#arguments-and-flags

Issue: https://github.com/omnibor/omnibor-rs/issues/243
Signed-off-by: Andrew Lilley Brinker <alilleybrinker@gmail.com>