feat(cli): impl 'store remove'
This implements the `store remove` command in the CLI, so users can now remove
manifests from the store if they want to. This required adding two calls to
the Storage trait and updating the in-memory and file-system storages. For
the FileSystemStorage type, I also opted to implement an interface similar to
the one used for upsert, and copied some of the code to ensure updated to the
target index are atomic filesystem operations. This code should probably be
pulled out into a reusable interface in the future.

Signed-off-by: Andrew Lilley Brinker <alilleybrinker@gmail.com>