feat(ffi): make `ArtifactIdSha256` type not opaque
Previously, the `ArtifactIdSha256` type was an opaque wrapper around the
Rust-side `ArtifactId<Sha256>`, but actually it's safe to transmute it to a
buffer of the appropriate size (the number of bytes for SHA-256 hash). This
commit now exposes the buffer to the C-side (still on the heap and shared as a
pointer, rather than on the stack to be copied, though this may change in the
future).

Signed-off-by: Andrew Lilley Brinker <alilleybrinker@gmail.com>