feat: Refactored embedding API.
This introduces a new `Embed` trait that is implemented by `NoEmbed` (which does
not do embedding), `AutoEmbed` (only if the `infer-filetypes` feature is turned
on, and which attempts to automatically infer filetypes for embedding), and by
any function of the appropriate signature, so users can substitute custom
embeddings.

This means without the `infer-filetypes` feature on, the supported options for
embedding are to do no embedding (via `NoEmbed`), or to provide your own. If
you provide your own, you likely know what type of file you're building, so
you can do without the heavy inference machinery.

Signed-off-by: Andrew Lilley Brinker <alilleybrinker@gmail.com>