feat: Introduced hash provider registry.
This is a major refactor that moves hash provider from being a frequent
argument needing to be passed in at call sites for many of the constructors in
this crate to being a global value.

This works by registering all of the hash providers turned on by compile-time
features, and then permitting the user to select an 'active' provider via a
global method.

Signed-off-by: Andrew Lilley Brinker <alilleybrinker@gmail.com>